requirements.txt
fastapi==0.100.0
openai==0.27.2
pgvector==0.1.2
asyncpg==0.27.0
pydantic==1.10.7
uvicorn==0.23.0
pytest==7.4.0
httpx==0.24.1
numpy~=1.26.4
python-dotenv~=1.0.1
psycopg2~=2.9.9
pdfplumber~=0.11.4
torch~=2.2.2
transformers~=4.44.2